requirements.txt
alembic==0.9.8
astroid==1.6.1
attrs==17.4.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
blinker==1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
coverage==4.0.3
coveralls==1.2.0
cryptography==1.7.2
dnspython==1.15.0
docopt==0.6.2
email-validator==1.0.2
enum34==1.1.6
flake8==3.5.0
flasgger==0.8.1
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-Validator==1.2.3
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.22
isbnlib==3.6.8
iso3166==0.8
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
passlib==1.7.1
pathlib==1.0.1
pew==1.1.2
pipenv==9.1.0
pluggy==0.6.0
psutil==5.3.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.2
py-moneyed==0.7.0
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.5.3
pylint==1.8.2
pylint-flask==0.5
pylint-plugin-utils==0.2.6
PyMySQL==0.8.0
pytest==3.4.0
pytest-cov==2.5.1
python-coveralls==2.9.1
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
requests==2.18.4
shutilwhich==1.1.0
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.2
urllib3==1.22
validate-email==1.3
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
